requirements.txt
beautifulsoup4==4.6.3
requests==2.19.1
slacker==0.9.65
tinydb==3.11.1
urllib3==1.23
pyyaml===3.13
